- Week day and week number are no longer clickable when onClick is missing for them respectively. - Remove references to ThemeColorField from old context theme. - SingleWeekSelection now highlights selected week using date range selection highlight.
