+++ packages/calendar/src/components/calendar/renderers/WeekNumberCell.tsx
@@ -21,30 +21,44 @@ export const WeekNumberCell: React.FC<WeekNumberCellProps> = ({
   backgroundColor,
   prefix,
 }) => {
+  const content = (
+    <Box
+      width={theme.width}
+      height={theme.height}
+      justifyContent={"center"}
+      alignItems={"center"}
+    >
+      {background && <Box position={"absolute"}>{background}</Box>}
+      <Box position={"absolute"}>
+        <StandardText
+          color={
+            onClickWeek
+              ? theme.WeekNumber.clickableTextColor
+              : theme.WeekNumber.textColor
+          }
+        >
+          {prefix}
+          {week.weekNumber}
+        </StandardText>
+      </Box>
+    </Box>
+  );
   return (
     <Box
       background={backgroundColor || theme.WeekNumber.backgroundColor}
       position={"relative"}
     >
-      <Clickable
-        onClick={onClickWeek ? (ev) => onClickWeek(week, ev) : undefined}
-        disableFocusHighlight={!onClickWeek}
-      >
-        <Box
-          width={theme.width}
-          height={theme.height}
-          justifyContent={"center"}
-          alignItems={"center"}
+      {onClickWeek ? (
+        <Clickable
+          borderRadius={"var(--swui-calendar-day-border-radius)"}
+          onClick={(ev) => onClickWeek(week, ev)}
+          disableFocusHighlight={!onClickWeek}
         >
-          {background && <Box position={"absolute"}>{background}</Box>}
-          <Box position={"absolute"}>
-            <StandardText color={theme.WeekNumber.textColor}>
-              {prefix}
-              {week.weekNumber}
-            </StandardText>
-          </Box>
-        </Box>
-      </Clickable>
+          {content}
+        </Clickable>
+      ) : (
+        content
+      )}
     </Box>
   );
 };
