+++ packages/calendar/src/components/input-types/date-range-input/DateRangeInput.tsx
@@ -17,9 +17,9 @@ import { useDateRangeInput } from "./hooks/UseDateRangeInput";
 import { Icon } from "@stenajs-webui/elements";
 import { faLongArrowAltRight } from "@fortawesome/free-solid-svg-icons/faLongArrowAltRight";
 import { CalendarWithMonthSwitcher } from "../../../features/month-switcher/CalendarWithMonthSwitcher";
-import { buildDayState } from "../../calendar-types/date-range-calendar/util/DayStateFactory";
 import { CalendarPanelType } from "../../../features/calendar-with-month-year-pickers/CalendarPanelType";
 import { Popover } from "@stenajs-webui/tooltip";
+import { buildDayStateForDateRange } from "../../../util/calendar/StateModifier";
 
 export interface DateRangeInputProps<T extends {}> {
   /** The current date range value */
@@ -120,7 +120,7 @@ export const DateRangeInput = <T extends {}>({
   useMultiOnClickOutside([popupRef, outsideRef], hideCalendar);
 
   const statePerMonth = useMemo(
-    () => buildDayState(undefined, value.startDate, value.endDate),
+    () => buildDayStateForDateRange(undefined, value.startDate, value.endDate),
     [value]
   );
